[:integtest:mysql] Rename MySQL routine test classes
