+++ integtest/mysql/src/test/java/io/github/zero88/integtest/jooqx/mysql/MySQLReAQueryTest.java
@@ -70,76 +70,4 @@ class MySQLReAQueryTest extends MySQLJooqxTest<MySQLPool> implements MySQLPoolPr
         })).onFailure(ctx::failNow);
     }
 
-    @Test
-    void test_fn_returns_value(VertxTestContext ctx) {
-        final Checkpoint cp = ctx.checkpoint();
-        final Hello routine = new Hello();
-        routine.setS("hi");
-        jooqx.routine(routine).onSuccess(out -> ctx.verify(() -> {
-            Assertions.assertEquals("Hello, hi!", out);
-            cp.flag();
-        })).onFailure(ctx::failNow);
-    }
-
-    @Test
-    @Disabled("Not yet supported out params in mysql reactive")
-    void test_proc_returns_outParams(VertxTestContext ctx) {
-        final Checkpoint cp = ctx.checkpoint();
-        final CountAuthorByCountry byCountry = new CountAuthorByCountry();
-        byCountry.setCountry("USA");
-        JooqSQL.printJooqRoutine(jooqx.dsl(), byCountry);
-
-        jooqx.sqlClient()
-             .getConnection()
-             .onFailure(ctx::failNow)
-             .flatMap(conn -> ((MySQLConnection) conn).setOption(MySQLSetOption.MYSQL_OPTION_MULTI_STATEMENTS_ON))
-             .onFailure(ctx::failNow)
-             .flatMap(unused -> jooqx.sqlQuery(
-                 dsl -> "call `test`.`count_author_by_country` ('USA', @count); SELECT @count;",
-                 DSLAdapter.fetchOne(DSL.field("count", String.class))))
-             .onSuccess(record -> ctx.verify(() -> {
-                 System.out.println(record);
-                 cp.flag();
-             }))
-             .onFailure(ctx::failNow);
-    }
-
-    @Test
-    void test_proc_returns_resultSet(VertxTestContext ctx) {
-        final Checkpoint cp = ctx.checkpoint();
-        final SelectBooksByAuthor proc = new SelectBooksByAuthor();
-        proc.setAuthor("Jane Austen");
-        JooqSQL.printJooqRoutine(jooqx.dsl(), proc);
-        jooqx.routineResultSet(proc, DSLAdapter.fetchMany(Books.BOOKS)).onSuccess(records -> ctx.verify(() -> {
-            Assertions.assertEquals(1, records.size());
-            BooksRecord rec = records.get(0);
-            Assertions.assertEquals(6, rec.getId());
-            Assertions.assertEquals("Pride and Prejudice", rec.getTitle());
-            cp.flag();
-        })).onFailure(ctx::failNow);
-    }
-
-    @Test
-    void test_proc_returns_void(VertxTestContext ctx) {
-        final Checkpoint cp = ctx.checkpoint(2);
-        final RemoveAuthor proc = new RemoveAuthor();
-        final String authorName = "J.D. Salinger";
-        proc.setAuthorName(authorName);
-        jooqx.routine(proc)
-             .flatMap(r -> jooqx.execute(dsl -> dsl.selectOne()
-                                                   .whereExists(dsl.selectFrom(Tables.AUTHORS)
-                                                                   .where(Tables.AUTHORS.NAME.eq(authorName))),
-                                         DSLAdapter.fetchExists()))
-             .onSuccess(r -> ctx.verify(() -> {
-                 Assertions.assertFalse(r);
-                 cp.flag();
-             }))
-             .flatMap(r -> jooqx.execute(dsl -> dsl.select().from(Tables.BOOKS), DSLAdapter.fetchMany(Tables.BOOKS)))
-             .onSuccess(records -> ctx.verify(() -> {
-                 Assertions.assertEquals(4, records.size());
-                 cp.flag();
-             }))
-             .onFailure(ctx::failNow);
-    }
-
 }
